python/web-utility/requirements.txt
djangorestframework
Django>=2.0,<3.0